Stripped down to just a raw library. This needs testing as the JS probably isn't valid yet.
