--- app/attestation.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>TAP - Transaction Attestation Platform</title>
-  <link href='https://fonts.googleapis.com/css?family=Open+Sans:400,700' rel='stylesheet' type='text/css'>
-  <link href="./app.css" rel='stylesheet' type='text/css'>
-  <script src="./ipfs.js"></script>
-  <script src="./app.js"></script>
-  <script src="https://unpkg.com/ipfs-api/dist/index.js"></script>
-</head>
-<body>
-  <h1>TAP - Transaction Attestation Platform</h1>
-  <h2>Attestation for <span id="contractName"></span>:<span id="methodId"></span></h2>
-
-  <h3>Contract level information</h3>
-  <br>
-  <div id="contractSummary"></div>
-
-  <br>
-  
-  <h3>Method level information</h3>
-  <br>
-  <div id="methodAttestation"></div>
-
-  <hr />
-  <span id="status"></span>
-</body>
-</html>
