--- app/stylesheets/app.css
@@ -1,81 +1,0 @@
-body {
-  margin-left: 25%;
-  margin-right: 25%;
-  margin-top: 10%;
-  font-family: "Open Sans", sans-serif;
-}
-
-label {
-  display: inline-block;
-  width: 100px;
-}
-
-input {
-  width: 500px;
-  padding: 5px;
-  font-size: 16px;
-}
-
-button {
-  font-size: 16px;
-  padding: 5px;
-}
-
-h1, h2 {
-  display: inline-block;
-  vertical-align: middle;
-  margin-top: 0px;
-  margin-bottom: 10px;
-}
-
-h2 {
-  color: #AAA;
-  font-size: 32px;
-}
-
-h3 {
-  font-weight: normal;
-  color: #AAA;
-  font-size: 24px;
-}
-
-.black {
-  color: black;
-}
-
-div#methodAttestation {
-    width: 800px;
-}
-
-#balance {
-  color: black;
-}
-
-#attestationTable {
-    width: 800px;
-    table-layout: fixed;
-}
-
-.trlabel {
-    width: 250px;
-    vertical-align: top;
-    font-weight: bold;
-}
-
-/* Add attestation page */
-.attCB {
-    width:20px;
-}
-
-.attLabel {
-    display: inline;
-}
-
-.attTextarea {
-    width:300px;
-    height:100px;
-}
-
-#exploitDescriptionDiv, #externalCallDescriptionDiv, #throwDescriptionDiv {
-    display:none;
-}
