--- app/verify.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>TAP - Transaction Attestation Platform</title>
-    <link href='https://fonts.googleapis.com/css?family=Open+Sans:400,700' rel='stylesheet' type='text/css'>
-    <link href="./app.css" rel='stylesheet' type='text/css'>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="./ipfs.js"></script>
-    <script src="https://unpkg.com/ipfs-api/dist/index.js"></script>
-    <script src="./app.js"></script>
-    <script src="./verify.js"></script>
-</head>
-<body>
-    <div id="compileError" style="color:red; display:none"></div>
-    <div style="margin:10px">
-        <label for="contractAddress">Address:</label><br>
-        <input type="text" id="contractAddress" style="width:500px"></input>
-    </div>
-    <div style="margin:10px; height:300px">
-        <label for="contractCode" >Code: </label><br>
-        <textarea id="contractCode" style="height:250px; width:508px; float:left"></textarea>
-        <button id="verifyButton" style="float:left; margin-top:228px; margin-left:20px">Verify</button>
-    </div>
-    <!-- <button id="compileButton">Compile</button> -->
-
-    <div id="compileResult" style="margin-top:20px;"></div>
-    <div id="contractBytecode" style="margin-top: 20px"></div>
-</body>
-</html>
